+++ traffic_ops/v3-client/deliveryservice_requests.go
@@ -16,11 +16,8 @@
 package client
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"net"
 	"net/http"
 	"net/url"
 
@@ -34,9 +31,8 @@ const (
 // CreateDeliveryServiceRequest creates a Delivery Service Request.
 func (to *Session) CreateDeliveryServiceRequest(dsr tc.DeliveryServiceRequest) (tc.Alerts, ReqInf, error) {
 	var alerts tc.Alerts
-	var remoteAddr net.Addr
 	if dsr.AssigneeID == 0 && dsr.Assignee != "" {
-		res, reqInf, err := to.GetUserByUsername(dsr.Assignee)
+		res, reqInf, err := to.GetUserByUsernameWithHdr(dsr.Assignee, nil)
 		if err != nil {
 			return alerts, reqInf, err
 		}
@@ -47,7 +43,7 @@ func (to *Session) CreateDeliveryServiceRequest(dsr tc.DeliveryServiceRequest) (
 	}
 
 	if dsr.AuthorID == 0 && dsr.Author != "" {
-		res, reqInf, err := to.GetUserByUsername(dsr.Author)
+		res, reqInf, err := to.GetUserByUsernameWithHdr(dsr.Author, nil)
 		if err != nil {
 			return alerts, reqInf, err
 		}
@@ -58,7 +54,7 @@ func (to *Session) CreateDeliveryServiceRequest(dsr tc.DeliveryServiceRequest) (
 	}
 
 	if dsr.DeliveryService.TypeID == 0 && dsr.DeliveryService.Type.String() != "" {
-		ty, reqInf, err := to.GetTypeByName(dsr.DeliveryService.Type.String())
+		ty, reqInf, err := to.GetTypeByNameWithHdr(dsr.DeliveryService.Type.String(), nil)
 		if err != nil || len(ty) == 0 {
 			return alerts, reqInf, errors.New("no type named " + dsr.DeliveryService.Type.String())
 		}
@@ -66,7 +62,7 @@ func (to *Session) CreateDeliveryServiceRequest(dsr tc.DeliveryServiceRequest) (
 	}
 
 	if dsr.DeliveryService.CDNID == 0 && dsr.DeliveryService.CDNName != "" {
-		cdns, reqInf, err := to.GetCDNByName(dsr.DeliveryService.CDNName)
+		cdns, reqInf, err := to.GetCDNByNameWithHdr(dsr.DeliveryService.CDNName, nil)
 		if err != nil || len(cdns) == 0 {
 			return alerts, reqInf, errors.New("no CDN named " + dsr.DeliveryService.CDNName)
 		}
@@ -74,7 +70,7 @@ func (to *Session) CreateDeliveryServiceRequest(dsr tc.DeliveryServiceRequest) (
 	}
 
 	if dsr.DeliveryService.ProfileID == 0 && dsr.DeliveryService.ProfileName != "" {
-		profiles, reqInf, err := to.GetProfileByName(dsr.DeliveryService.ProfileName)
+		profiles, reqInf, err := to.GetProfileByNameWithHdr(dsr.DeliveryService.ProfileName, nil)
 		if err != nil || len(profiles) == 0 {
 			return alerts, reqInf, errors.New("no Profile named " + dsr.DeliveryService.ProfileName)
 		}
@@ -82,57 +78,23 @@ func (to *Session) CreateDeliveryServiceRequest(dsr tc.DeliveryServiceRequest) (
 	}
 
 	if dsr.DeliveryService.TenantID == 0 && dsr.DeliveryService.Tenant != "" {
-		ten, reqInf, err := to.TenantByName(dsr.DeliveryService.Tenant)
+		ten, reqInf, err := to.TenantByNameWithHdr(dsr.DeliveryService.Tenant, nil)
 		if err != nil || ten == nil {
 			return alerts, reqInf, errors.New("no Tenant named " + dsr.DeliveryService.Tenant)
 		}
 		dsr.DeliveryService.TenantID = ten.ID
 	}
 
-	reqBody, err := json.Marshal(dsr)
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if err != nil {
-		return alerts, reqInf, err
-	}
-	resp, remoteAddr, err := to.RawRequest(http.MethodPost, API_DS_REQUESTS, reqBody)
-	defer resp.Body.Close()
-
-	if err == nil {
-		body, readErr := ioutil.ReadAll(resp.Body)
-		if readErr != nil {
-			return alerts, reqInf, readErr
-		}
-		if err = json.Unmarshal(body, &alerts); err != nil {
-			return alerts, reqInf, err
-		}
-	}
-
+	reqInf, err := to.post(API_DS_REQUESTS, dsr, nil, &alerts)
 	return alerts, reqInf, err
 }
 
 func (to *Session) GetDeliveryServiceRequestsWithHdr(header http.Header) ([]tc.DeliveryServiceRequest, ReqInf, error) {
-	resp, remoteAddr, err := to.request(http.MethodGet, API_DS_REQUESTS, nil, header)
-
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if resp != nil {
-		reqInf.StatusCode = resp.StatusCode
-		if reqInf.StatusCode == http.StatusNotModified {
-			return []tc.DeliveryServiceRequest{}, reqInf, nil
-		}
-	}
-	if err != nil {
-		return nil, reqInf, err
-	}
-	defer resp.Body.Close()
-
 	data := struct {
 		Response []tc.DeliveryServiceRequest `json:"response"`
 	}{}
-	if err = json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return nil, reqInf, err
-	}
-
-	return data.Response, reqInf, nil
+	reqInf, err := to.get(API_DS_REQUESTS, header, &data)
+	return data.Response, reqInf, err
 }
 
 // GetDeliveryServiceRequests retrieves all deliveryservices available to session user.
@@ -143,28 +105,11 @@ func (to *Session) GetDeliveryServiceRequests() ([]tc.DeliveryServiceRequest, Re
 
 func (to *Session) GetDeliveryServiceRequestByXMLIDWithHdr(XMLID string, header http.Header) ([]tc.DeliveryServiceRequest, ReqInf, error) {
 	route := fmt.Sprintf("%s?xmlId=%s", API_DS_REQUESTS, url.QueryEscape(XMLID))
-	resp, remoteAddr, err := to.request(http.MethodGet, route, nil, header)
-
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if resp != nil {
-		reqInf.StatusCode = resp.StatusCode
-		if reqInf.StatusCode == http.StatusNotModified {
-			return []tc.DeliveryServiceRequest{}, reqInf, nil
-		}
-	}
-	if err != nil {
-		return nil, reqInf, err
-	}
-	defer resp.Body.Close()
-
 	data := struct {
 		Response []tc.DeliveryServiceRequest `json:"response"`
 	}{}
-	if err = json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return nil, reqInf, err
-	}
-
-	return data.Response, reqInf, nil
+	reqInf, err := to.get(route, header, &data)
+	return data.Response, reqInf, err
 }
 
 // GET a DeliveryServiceRequest by the DeliveryServiceRequest XMLID
@@ -175,27 +120,11 @@ func (to *Session) GetDeliveryServiceRequestByXMLID(XMLID string) ([]tc.Delivery
 
 func (to *Session) GetDeliveryServiceRequestByIDWithHdr(id int, header http.Header) ([]tc.DeliveryServiceRequest, ReqInf, error) {
 	route := fmt.Sprintf("%s?id=%d", API_DS_REQUESTS, id)
-	resp, remoteAddr, err := to.request(http.MethodGet, route, nil, header)
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if resp != nil {
-		reqInf.StatusCode = resp.StatusCode
-		if reqInf.StatusCode == http.StatusNotModified {
-			return []tc.DeliveryServiceRequest{}, reqInf, nil
-		}
-	}
-	if err != nil {
-		return nil, reqInf, err
-	}
-	defer resp.Body.Close()
-
 	data := struct {
 		Response []tc.DeliveryServiceRequest `json:"response"`
 	}{}
-	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return nil, reqInf, err
-	}
-
-	return data.Response, reqInf, nil
+	reqInf, err := to.get(route, header, &data)
+	return data.Response, reqInf, err
 }
 
 // GET a DeliveryServiceRequest by the DeliveryServiceRequest id
@@ -206,24 +135,9 @@ func (to *Session) GetDeliveryServiceRequestByID(id int) ([]tc.DeliveryServiceRe
 
 func (to *Session) UpdateDeliveryServiceRequestByIDWithHdr(id int, dsr tc.DeliveryServiceRequest, header http.Header) (tc.Alerts, ReqInf, error) {
 	route := fmt.Sprintf("%s?id=%d", API_DS_REQUESTS, id)
-
-	var remoteAddr net.Addr
-	reqBody, err := json.Marshal(dsr)
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if err != nil {
-		return tc.Alerts{}, reqInf, err
-	}
-	resp, remoteAddr, err := to.request(http.MethodPut, route, reqBody, header)
-	if resp != nil {
-		reqInf.StatusCode = resp.StatusCode
-	}
-	if err != nil {
-		return tc.Alerts{}, reqInf, err
-	}
-	defer resp.Body.Close()
 	var alerts tc.Alerts
-	err = json.NewDecoder(resp.Body).Decode(&alerts)
-	return alerts, reqInf, nil
+	reqInf, err := to.put(route, dsr, header, &alerts)
+	return alerts, reqInf, err
 }
 
 // Update a DeliveryServiceRequest by ID
@@ -235,49 +149,7 @@ func (to *Session) UpdateDeliveryServiceRequestByID(id int, dsr tc.DeliveryServi
 // DELETE a DeliveryServiceRequest by DeliveryServiceRequest assignee
 func (to *Session) DeleteDeliveryServiceRequestByID(id int) (tc.Alerts, ReqInf, error) {
 	route := fmt.Sprintf("%s?id=%d", API_DS_REQUESTS, id)
-	resp, remoteAddr, err := to.RawRequest(http.MethodDelete, route, nil)
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if err != nil {
-		return tc.Alerts{}, reqInf, err
-	}
-	defer resp.Body.Close()
 	var alerts tc.Alerts
-	err = json.NewDecoder(resp.Body).Decode(&alerts)
-	return alerts, reqInf, nil
+	reqInf, err := to.del(route, nil, &alerts)
+	return alerts, reqInf, err
 }
-
-/*
-
-// Returns a list of DeliveryServiceRequests
-func (to *Session) GetDeliveryServiceRequests() ([]tc.DeliveryServiceRequest, ReqInf, error) {
-	resp, remoteAddr, err := to.request(http.MethodGet, API_DS_REQUESTS, nil)
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if err != nil {
-		return nil, reqInf, err
-	}
-	defer resp.Body.Close()
-
-	var data tc.DeliveryServiceRequestsResponse
-	err = json.NewDecoder(resp.Body).Decode(&data)
-	return data.Response, reqInf, nil
-}
-
-// GET a DeliveryServiceRequest by the DeliveryServiceRequest assignee
-func (to *Session) GetDeliveryServiceRequestByAssignee(assignee string) ([]tc.DeliveryServiceRequest, ReqInf, error) {
-	url := fmt.Sprintf("%s/assignee/%s", API_DS_REQUESTS, assignee)
-	resp, remoteAddr, err := to.request(http.MethodGet, url, nil)
-	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
-	if err != nil {
-		return nil, reqInf, err
-	}
-	defer resp.Body.Close()
-
-	var data tc.DeliveryServiceRequestsResponse
-	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return nil, reqInf, err
-	}
-
-	return data.Response, reqInf, nil
-}
-
-*/